- spliting movies details stack into smaller widgets.
