+++ lib/features/movies/ui/widgets/movie_details/movie_details_success.dart
@@ -1,11 +1,11 @@
-import '../../../../../core/helpers/api_data_helper.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
 import '../../../data/models/movie_details_model.dart';
 import '../../../logic/movies_details_cubit/movies_details_cubit.dart';
 import 'cast_and_crew.dart';
+import 'movie_details_stack/movies_details_stack.dart';
 import 'movies_details_overview.dart';
-import 'movies_details_stack.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 class MovieDetailsSuccess extends StatelessWidget {
   const MovieDetailsSuccess({super.key, required this.movieDetails});
@@ -13,15 +13,9 @@ class MovieDetailsSuccess extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final double screenHeight = MediaQuery.of(context).size.height;
-    final double screenWidth = MediaQuery.of(context).size.width;
-    String imageUrl = ApiDataHelper.getImageUrl(path: movieDetails.posterPath!);
     return ListView(
       children: [
         MoviesDetailsStack(
-          imageUrl: imageUrl,
-          screenHeight: screenHeight,
-          screenWidth: screenWidth,
           movieDetails: movieDetails,
         ),
         MoviesDetailsOverview(overview: movieDetails.overview!),
